Prefer u64 for num_descendants

+++ src/syntax.rs
@@ -53,9 +53,7 @@ pub enum Syntax<'a> {
         children: Vec<&'a Syntax<'a>>,
         close_position: Vec<SingleLineSpan>,
         close_content: String,
-        // TODO: this probably makes more sense as a u64, since we
-        // don't use it for indexing.
-        num_descendants: usize,
+        num_descendants: u64,
     },
     Atom {
         info: SyntaxInfo<'a>,
